Fixed a concept base to work with
Concept base will be used primarily to work with the whole of the data
and create graphs to that. adg.hpp and adg.cpp include simple functions
for abstractions and thread safe (hopefully)

Ensured leaf nodes exists and changed CMakeLists.txt to work with the
current filesystem. The problem of ccls not recognizing the includes
directories still remains to work with.